Refactor DoFnOp.FutureCollectorImpl to a top level class in SamzaRunner
