+++ runners/samza/src/test/java/org/apache/beam/runners/samza/runtime/FutureCollectorImplTest.java
@@ -31,14 +31,14 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-/** Unit tests for {@linkplain org.apache.beam.runners.samza.runtime.DoFnOp.FutureCollectorImpl}. */
+/** Unit tests for {@linkplain FutureCollectorImpl}. */
 public final class FutureCollectorImplTest {
   private static final List<String> RESULTS = ImmutableList.of("hello", "world");
-  private FutureCollector<String> futureCollector = new DoFnOp.FutureCollectorImpl<>();
+  private FutureCollector<String> futureCollector = new FutureCollectorImpl<>();
 
   @Before
   public void setup() {
-    futureCollector = new DoFnOp.FutureCollectorImpl<>();
+    futureCollector = new FutureCollectorImpl<>();
   }
 
   @Test(expected = IllegalStateException.class)
